fixed all leak and memory problem
except leak pcap find all devs of course

+++ Progetto/2parte/Receiver/ImagePicker/ImagePicker/receiver.h
@@ -22,7 +22,7 @@ void create_header_information(u_char* rtpdata, unsigned int fragment_type, int 
 
 //aggiungi i dati di un rtp al payload
 //ritorna uno se il gop è terminato 0 altrimenti
-int addPacketToGOP(u_char* rtpdata, int rtpdata_len, int seq_num, rtp* el, gop_info* info);
+int addPacketToGOP(u_char* rtpdata, int rtpdata_len, uint16_t seq_num, rtp* el, gop_info* info);
 
 //stub per chiamare addPacket e e create_header in autonomia
 //ritorna uno se il gop è terminato 0 altrimenti
@@ -32,7 +32,7 @@ int workOnPacket(rtp* el, gop_info* info);
 void save_GOP(int *from, gop_info* info);
 
 //funzioni di utlità per la tabella hash
-void* insert_hash(int primarykey, void* insert);
+void* insert_hash(uint16_t primarykey, void* insert);
 
 void* find_hash(int* primarykey);
